move device under MokiRuby
