--- lib/device.rb
@@ -1,32 +1,0 @@
-class Device
-  attr :id, :identifier_type
-
-  def initialize(identifier)
-    if is_serial?(identifier)
-      @identifier_type = :serial
-    elsif is_udid?(identifier)
-      @identifier_type = :udid
-    else
-      raise "Valid UDID or Serial Number required"
-    end
-    @id = identifier
-  end
-
-  def device_id_param
-    if identifier_type == :serial
-      "sn-!-#{ id }"
-    else
-      id
-    end
-  end
-
-  private
-
-  def is_serial?(id)
-    (!id.nil? && id.length == 12)
-  end
-
-  def is_udid?(id)
-    !(/\A[a-f0-9]{8}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{12}\Z/.match(id)).nil?
-  end
-end
