+++ spec/lib/device_spec.rb
@@ -1,14 +1,14 @@
 require 'spec_helper'
-require 'device'
+require 'moki_ruby/device'
 
-describe Device do
+describe MokiRuby::Device do
   let(:sn) { "ABCDEFGHIJ12" }
   let(:udid) { "abcd1234-1234-1234-1234-abcdef123456" }
 
   describe 'initialization' do
     context "with serial number" do
       it "is valid" do
-        device = Device.new(sn)
+        device = MokiRuby::Device.new(sn)
         expect(device.id).to eq sn
         expect(device.identifier_type).to eq :serial
       end
@@ -16,14 +16,14 @@ describe Device do
 
     context "with UDID" do
       it "is valid" do
-        device = Device.new(udid)
+        device = MokiRuby::Device.new(udid)
         expect(device.id).to eq udid
       end
     end
 
     context "with invalid identifier" do
       it "is invalid" do
-        expect{ Device.new() }.to raise_error
+        expect{ MokiRuby::Device.new() }.to raise_error
       end
     end
   end
@@ -31,14 +31,14 @@ describe Device do
   describe "#device_id_param" do
     describe "Serial Number" do
       it "prepends serial number data" do
-        device = Device.new(sn)
+        device = MokiRuby::Device.new(sn)
         expect(device.device_id_param).to eq "sn-!-#{ sn }"
       end
     end
 
     describe "UDID" do
       it "prepends serial number data" do
-        device = Device.new(udid)
+        device = MokiRuby::Device.new(udid)
         expect(device.device_id_param).to eq udid
       end
     end
